improve & unify serialization tests for triggers
