--- tests/providers/amazon/aws/triggers/test_redshift_cluster.py
@@ -1,81 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-from __future__ import annotations
-
-import pytest
-
-from airflow.providers.amazon.aws.triggers.redshift_cluster import (
-    RedshiftCreateClusterSnapshotTrigger,
-    RedshiftCreateClusterTrigger,
-    RedshiftDeleteClusterTrigger,
-    RedshiftPauseClusterTrigger,
-    RedshiftResumeClusterTrigger,
-)
-
-TEST_CLUSTER_IDENTIFIER = "test-cluster"
-TEST_POLL_INTERVAL = 10
-TEST_MAX_ATTEMPT = 10
-TEST_AWS_CONN_ID = "test-aws-id"
-
-
-class TestRedshiftClusterTriggers:
-    @pytest.mark.parametrize(
-        "trigger",
-        [
-            RedshiftCreateClusterTrigger(
-                cluster_identifier=TEST_CLUSTER_IDENTIFIER,
-                poll_interval=TEST_POLL_INTERVAL,
-                max_attempt=TEST_MAX_ATTEMPT,
-                aws_conn_id=TEST_AWS_CONN_ID,
-            ),
-            RedshiftPauseClusterTrigger(
-                cluster_identifier=TEST_CLUSTER_IDENTIFIER,
-                poll_interval=TEST_POLL_INTERVAL,
-                max_attempts=TEST_MAX_ATTEMPT,
-                aws_conn_id=TEST_AWS_CONN_ID,
-            ),
-            RedshiftCreateClusterSnapshotTrigger(
-                cluster_identifier=TEST_CLUSTER_IDENTIFIER,
-                poll_interval=TEST_POLL_INTERVAL,
-                max_attempts=TEST_MAX_ATTEMPT,
-                aws_conn_id=TEST_AWS_CONN_ID,
-            ),
-            RedshiftResumeClusterTrigger(
-                cluster_identifier=TEST_CLUSTER_IDENTIFIER,
-                poll_interval=TEST_POLL_INTERVAL,
-                max_attempts=TEST_MAX_ATTEMPT,
-                aws_conn_id=TEST_AWS_CONN_ID,
-            ),
-            RedshiftDeleteClusterTrigger(
-                cluster_identifier=TEST_CLUSTER_IDENTIFIER,
-                poll_interval=TEST_POLL_INTERVAL,
-                max_attempts=TEST_MAX_ATTEMPT,
-                aws_conn_id=TEST_AWS_CONN_ID,
-            ),
-        ],
-    )
-    def test_serialize_recreate(self, trigger):
-        class_path, args = trigger.serialize()
-
-        class_name = class_path.split(".")[-1]
-        clazz = globals()[class_name]
-        instance = clazz(**args)
-
-        class_path2, args2 = instance.serialize()
-
-        assert class_path == class_path2
-        assert args == args2
