+++ tests/providers/amazon/aws/triggers/test_sqs.py
@@ -50,24 +50,6 @@ trigger = SqsSensorTrigger(
 
 
 class TestSqsTriggers:
-    @pytest.mark.parametrize(
-        "trigger",
-        [
-            trigger,
-        ],
-    )
-    def test_serialize_recreate(self, trigger):
-        class_path, args = trigger.serialize()
-
-        class_name = class_path.split(".")[-1]
-        clazz = globals()[class_name]
-        instance = clazz(**args)
-
-        class_path2, args2 = instance.serialize()
-
-        assert class_path == class_path2
-        assert args == args2
-
     @pytest.mark.asyncio
     async def test_poke(self):
         sqs_trigger = trigger
